Move anti-logger to main app class
