+++ server/src/main/java/org/mskcc/smile/SmileConfiguration.java
@@ -1,10 +1,6 @@
 package org.mskcc.smile;
 
-import ch.qos.logback.classic.Level;
-import ch.qos.logback.classic.Logger;
 import org.neo4j.ogm.session.SessionFactory;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -42,22 +38,5 @@ public class SmileConfiguration {
                 .uri(uri)
                 .credentials(username, password)
                 .build();
-    }
-
-    /**
-     * Added this as a means of silencing the neo4j ogm log cluttering.
-     * There's probably a nicer way to do this but it gets the job done.
-     */
-    @Autowired
-    public void logger() {
-        Logger logger = (Logger)
-                LoggerFactory.getLogger("org.neo4j.ogm.drivers.bolt.response.BoltResponse.unrecognized");
-        logger.setLevel(Level.OFF);
-        logger = (Logger)
-                LoggerFactory.getLogger("org.neo4j.ogm.context.GraphEntityMapper");
-        logger.setLevel(Level.OFF);
-        logger = (Logger)
-                LoggerFactory.getLogger("org.neo4j.ogm.context.EntityGraphMapper");
-        logger.setLevel(Level.OFF);
     }
 }
